Split hooks into separate files
