+++ public/app/features/provisioning/hooks/index.ts
@@ -1,0 +1,6 @@
+export { useCreateOrUpdateRepository } from './useCreateOrUpdateRepository';
+export { useCreateOrUpdateRepositoryFile } from './useCreateOrUpdateRepositoryFile';
+export { useGetResourceRepository } from './useGetResourceRepository';
+export { usePullRequestParam } from './usePullRequestParam';
+export { useRepositoryJobs } from './useRepositoryJobs';
+export { useRepositoryList } from './useRepositoryList';
